Cut out the Table of Times as TimesTable #1
